Revision: 201017 Kit: 201019

+++ homescreenapp/hsapplication/conf/base/implml/homescreendb_templates/create.sql
@@ -5,10 +5,10 @@ CREATE TABLE Scene (
   landscapeWallpaper TEXT,
   defaultPageId INTEGER,
   maximumPageCount INTEGER,
-  maximumWidgetHeight INTEGER,
-  maximumWidgetWidth INTEGER,
-  minimumWidgetHeight INTEGER,
-  minimumWidgetWidth INTEGER);
+  maximumWidgetHeight REAL,
+  maximumWidgetWidth REAL,
+  minimumWidgetHeight REAL,
+  minimumWidgetWidth REAL);
 
 DROP TABLE IF EXISTS Pages;
 CREATE TABLE Pages (
